% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{authorize_snapshot_access}
\alias{authorize_snapshot_access}
\title{Authorizes the specified AWS customer account to restore the specified snapshot}
\usage{
authorize_snapshot_access(SnapshotIdentifier, AccountWithRestoreAccess,
  SnapshotClusterIdentifier = NULL)
}
\arguments{
\item{SnapshotIdentifier}{The identifier of the snapshot the account is authorized to restore.}

\item{AccountWithRestoreAccess}{The identifier of the AWS customer account authorized to restore the specified snapshot.

To share a snapshot with AWS support, specify amazon-redshift-support.}

\item{SnapshotClusterIdentifier}{The identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.}
}
\description{
Authorizes the specified AWS customer account to restore the specified snapshot.
}
\details{
For more information about working with snapshots, go to \href{http://docs.aws.amazon.com/redshift/latest/mgmt/working-with-snapshots.html}{Amazon Redshift Snapshots} in the \emph{Amazon Redshift Cluster Management Guide}.
}
\examples{

}
