% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{batch_delete_cluster_snapshots}
\alias{batch_delete_cluster_snapshots}
\title{Deletes a set of cluster snapshots}
\usage{
batch_delete_cluster_snapshots(Identifiers)
}
\arguments{
\item{Identifiers}{[required] A list of identifiers for the snapshots that you want to delete.}
}
\description{
Deletes a set of cluster snapshots.
}
\section{Accepted Parameters}{
\preformatted{batch_delete_cluster_snapshots(
  Identifiers = list(
    list(
      SnapshotIdentifier = "string",
      SnapshotClusterIdentifier = "string"
    )
  )
)
}
}

