% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{cancel_resize}
\alias{cancel_resize}
\title{Cancels a resize operation}
\usage{
cancel_resize(ClusterIdentifier)
}
\arguments{
\item{ClusterIdentifier}{[required] The unique identifier for the cluster that you want to cancel a resize operation for.}
}
\description{
Cancels a resize operation.
}
\section{Accepted Parameters}{
\preformatted{cancel_resize(
  ClusterIdentifier = "string"
)
}
}

