% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{copy_cluster_snapshot}
\alias{copy_cluster_snapshot}
\title{Copies the specified automated cluster snapshot to a new manual cluster snapshot}
\usage{
copy_cluster_snapshot(
  SourceSnapshotIdentifier = "string",
  SourceSnapshotClusterIdentifier = "string",
  TargetSnapshotIdentifier = "string",
  ManualSnapshotRetentionPeriod = 123
)
}
\arguments{
\item{SourceSnapshotIdentifier}{[required] The identifier for the source snapshot.

Constraints:
\itemize{
\item Must be the identifier for a valid automated snapshot whose state is \code{available}.
}}

\item{SourceSnapshotClusterIdentifier}{The identifier of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

Constraints:
\itemize{
\item Must be the identifier for a valid cluster.
}}

\item{TargetSnapshotIdentifier}{[required] The identifier given to the new manual snapshot.

Constraints:
\itemize{
\item Cannot be null, empty, or blank.
\item Must contain from 1 to 255 alphanumeric characters or hyphens.
\item First character must be a letter.
\item Cannot end with a hyphen or contain two consecutive hyphens.
\item Must be unique for the AWS account that is making the request.
}}

\item{ManualSnapshotRetentionPeriod}{The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

The default value is -1.}
}
\description{
Copies the specified automated cluster snapshot to a new manual cluster snapshot. The source must be an automated snapshot and it must be in the available state.
}
\details{
When you delete a cluster, Amazon Redshift deletes any automated snapshots of the cluster. Also, when the retention period of the snapshot expires, Amazon Redshift automatically deletes it. If you want to keep an automated snapshot for a longer period, you can make a manual copy of the snapshot. Manual snapshots are retained until you delete them.

For more information about working with snapshots, go to \href{http://docs.aws.amazon.com/redshift/latest/mgmt/working-with-snapshots.html}{Amazon Redshift Snapshots} in the \emph{Amazon Redshift Cluster Management Guide}.
}
\examples{

}
