% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{create_cluster_snapshot}
\alias{create_cluster_snapshot}
\title{Creates a manual snapshot of the specified cluster}
\usage{
create_cluster_snapshot(SnapshotIdentifier, ClusterIdentifier, Tags = NULL)
}
\arguments{
\item{SnapshotIdentifier}{A unique identifier for the snapshot that you are requesting. This identifier must be unique for all snapshots within the AWS account.

Constraints:
\itemize{
\item Cannot be null, empty, or blank
\item Must contain from 1 to 255 alphanumeric characters or hyphens
\item First character must be a letter
\item Cannot end with a hyphen or contain two consecutive hyphens
}

Example: \code{my-snapshot-id}}

\item{ClusterIdentifier}{The cluster identifier for which you want a snapshot.}

\item{Tags}{A list of tag instances.}
}
\description{
Creates a manual snapshot of the specified cluster. The cluster must be in the \code{available} state.
}
\details{
For more information about working with snapshots, go to \href{http://docs.aws.amazon.com/redshift/latest/mgmt/working-with-snapshots.html}{Amazon Redshift Snapshots} in the \emph{Amazon Redshift Cluster Management Guide}.
}
\examples{

}
