% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{delete_cluster}
\alias{delete_cluster}
\title{Deletes a previously provisioned cluster}
\usage{
delete_cluster(ClusterIdentifier, SkipFinalClusterSnapshot = NULL,
  FinalClusterSnapshotIdentifier = NULL)
}
\arguments{
\item{ClusterIdentifier}{The identifier of the cluster to be deleted.

Constraints:
\itemize{
\item Must contain lowercase characters.
\item Must contain from 1 to 63 alphanumeric characters or hyphens.
\item First character must be a letter.
\item Cannot end with a hyphen or contain two consecutive hyphens.
}}

\item{SkipFinalClusterSnapshot}{Determines whether a final snapshot of the cluster is created before Amazon Redshift deletes the cluster. If \code{true}, a final cluster snapshot is not created. If \code{false}, a final cluster snapshot is created before the cluster is deleted.

The \emph{FinalClusterSnapshotIdentifier} parameter must be specified if \emph{SkipFinalClusterSnapshot} is \code{false}.

Default: \code{false}}

\item{FinalClusterSnapshotIdentifier}{The identifier of the final snapshot that is to be created immediately before deleting the cluster. If this parameter is provided, \emph{SkipFinalClusterSnapshot} must be \code{false}.

Constraints:
\itemize{
\item Must be 1 to 255 alphanumeric characters.
\item First character must be a letter.
\item Cannot end with a hyphen or contain two consecutive hyphens.
}}
}
\description{
Deletes a previously provisioned cluster. A successful response from the web service indicates that the request was received correctly. Use DescribeClusters to monitor the status of the deletion. The delete operation cannot be canceled or reverted once submitted. For more information about managing clusters, go to \href{http://docs.aws.amazon.com/redshift/latest/mgmt/working-with-clusters.html}{Amazon Redshift Clusters} in the \emph{Amazon Redshift Cluster Management Guide}.
}
\details{
If you want to shut down the cluster and retain it for future use, set \emph{SkipFinalClusterSnapshot} to \code{false} and specify a name for \emph{FinalClusterSnapshotIdentifier}. You can later restore this snapshot to resume using the cluster. If a final cluster snapshot is requested, the status of the cluster will be "final-snapshot" while the snapshot is being taken, then it's "deleting" once Amazon Redshift begins deleting the cluster.

For more information about managing clusters, go to \href{http://docs.aws.amazon.com/redshift/latest/mgmt/working-with-clusters.html}{Amazon Redshift Clusters} in the \emph{Amazon Redshift Cluster Management Guide}.
}
\examples{

}
