% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{delete_cluster_security_group}
\alias{delete_cluster_security_group}
\title{Deletes an Amazon Redshift security group}
\usage{
delete_cluster_security_group(ClusterSecurityGroupName)
}
\arguments{
\item{ClusterSecurityGroupName}{[required] The name of the cluster security group to be deleted.}
}
\description{
Deletes an Amazon Redshift security group.
}
\details{
You cannot delete a security group that is associated with any clusters. You cannot delete the default security group.

For information about managing security groups, go to \href{http://docs.aws.amazon.com/redshift/latest/mgmt/working-with-security-groups.html}{Amazon Redshift Cluster Security Groups} in the \emph{Amazon Redshift Cluster Management Guide}.
}
\section{Accepted Parameters}{
\preformatted{delete_cluster_security_group(
  ClusterSecurityGroupName = "string"
)
}
}

