% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{delete_cluster_subnet_group}
\alias{delete_cluster_subnet_group}
\title{Deletes the specified cluster subnet group}
\usage{
delete_cluster_subnet_group(ClusterSubnetGroupName)
}
\arguments{
\item{ClusterSubnetGroupName}{The name of the cluster subnet group name to be deleted.}
}
\description{
Deletes the specified cluster subnet group.
}
\examples{

}
