% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{delete_event_subscription}
\alias{delete_event_subscription}
\title{Deletes an Amazon Redshift event notification subscription}
\usage{
delete_event_subscription(
  SubscriptionName = "string"
)
}
\arguments{
\item{SubscriptionName}{[required] The name of the Amazon Redshift event notification subscription to be deleted.}
}
\description{
Deletes an Amazon Redshift event notification subscription.
}
\examples{

}
