% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{delete_hsm_client_certificate}
\alias{delete_hsm_client_certificate}
\title{Deletes the specified HSM client certificate}
\usage{
delete_hsm_client_certificate(
  HsmClientCertificateIdentifier = "string"
)
}
\arguments{
\item{HsmClientCertificateIdentifier}{[required] The identifier of the HSM client certificate to be deleted.}
}
\description{
Deletes the specified HSM client certificate.
}
\examples{

}
