% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{delete_hsm_configuration}
\alias{delete_hsm_configuration}
\title{Deletes the specified Amazon Redshift HSM configuration}
\usage{
delete_hsm_configuration(HsmConfigurationIdentifier)
}
\arguments{
\item{HsmConfigurationIdentifier}{[required] The identifier of the Amazon Redshift HSM configuration to be deleted.}
}
\description{
Deletes the specified Amazon Redshift HSM configuration.
}
\section{Accepted Parameters}{
\preformatted{delete_hsm_configuration(
  HsmConfigurationIdentifier = "string"
)
}
}

