% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{delete_snapshot_schedule}
\alias{delete_snapshot_schedule}
\title{Deletes a snapshot schedule}
\usage{
delete_snapshot_schedule(
  ScheduleIdentifier = "string"
)
}
\arguments{
\item{ScheduleIdentifier}{[required] A unique identifier of the snapshot schedule to delete.}
}
\description{
Deletes a snapshot schedule.
}
\examples{

}
