% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{describe_account_attributes}
\alias{describe_account_attributes}
\title{Returns a list of attributes attached to an account}
\usage{
describe_account_attributes(
  AttributeNames = list(
    "string"
  )
)
}
\arguments{
\item{AttributeNames}{A list of attribute names.}
}
\description{
Returns a list of attributes attached to an account
}
\examples{

}
