% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{describe_event_categories}
\alias{describe_event_categories}
\title{Displays a list of event categories for all event source types, or for a specified source type}
\usage{
describe_event_categories(SourceType = NULL)
}
\arguments{
\item{SourceType}{The source type, such as cluster or parameter group, to which the described event categories apply.

Valid values: cluster, cluster-snapshot, cluster-parameter-group, and cluster-security-group.}
}
\description{
Displays a list of event categories for all event source types, or for a specified source type. For a list of the event categories and source types, go to \href{http://docs.aws.amazon.com/redshift/latest/mgmt/working-with-event-notifications.html}{Amazon Redshift Event Notifications}.
}
\examples{

}
