% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{describe_logging_status}
\alias{describe_logging_status}
\title{Describes whether information, such as queries and connection attempts, is being logged for the specified Amazon Redshift cluster}
\usage{
describe_logging_status(ClusterIdentifier)
}
\arguments{
\item{ClusterIdentifier}{[required] The identifier of the cluster from which to get the logging status.

Example: \code{examplecluster}}
}
\description{
Describes whether information, such as queries and connection attempts, is being logged for the specified Amazon Redshift cluster.
}
\section{Accepted Parameters}{
\preformatted{describe_logging_status(
  ClusterIdentifier = "string"
)
}
}

