% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{describe_storage}
\alias{describe_storage}
\title{Returns the total amount of snapshot usage and provisioned storage for a user in megabytes}
\usage{
describe_storage()
}
\description{
Returns the total amount of snapshot usage and provisioned storage for a user in megabytes.
}
