% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{disable_logging}
\alias{disable_logging}
\title{Stops logging information, such as queries and connection attempts, for the specified Amazon Redshift cluster}
\usage{
disable_logging(
  ClusterIdentifier = "string"
)
}
\arguments{
\item{ClusterIdentifier}{[required] The identifier of the cluster on which logging is to be stopped.

Example: \code{examplecluster}}
}
\description{
Stops logging information, such as queries and connection attempts, for the specified Amazon Redshift cluster.
}
\examples{

}
