% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{enable_snapshot_copy}
\alias{enable_snapshot_copy}
\title{Enables the automatic copy of snapshots from one region to another region for a specified cluster}
\usage{
enable_snapshot_copy(ClusterIdentifier, DestinationRegion,
  RetentionPeriod = NULL, SnapshotCopyGrantName = NULL)
}
\arguments{
\item{ClusterIdentifier}{The unique identifier of the source cluster to copy snapshots from.

Constraints: Must be the valid name of an existing cluster that does not already have cross-region snapshot copy enabled.}

\item{DestinationRegion}{The destination region that you want to copy snapshots to.

Constraints: Must be the name of a valid region. For more information, see \href{http://docs.aws.amazon.com/general/latest/gr/rande.html#redshift_region}{Regions and Endpoints} in the Amazon Web Services General Reference.}

\item{RetentionPeriod}{The number of days to retain automated snapshots in the destination region after they are copied from the source region.

Default: 7.

Constraints: Must be at least 1 and no more than 35.}

\item{SnapshotCopyGrantName}{The name of the snapshot copy grant to use when snapshots of an AWS KMS-encrypted cluster are copied to the destination region.}
}
\description{
Enables the automatic copy of snapshots from one region to another region for a specified cluster.
}
\examples{

}
