% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{modify_cluster_db_revision}
\alias{modify_cluster_db_revision}
\title{Modifies the database revision of a cluster}
\usage{
modify_cluster_db_revision(ClusterIdentifier, RevisionTarget)
}
\arguments{
\item{ClusterIdentifier}{The unique identifier of a cluster whose database revision you want to modify.

Example: \code{examplecluster}}

\item{RevisionTarget}{The identifier of the database revision. You can retrieve this value from the response to the DescribeClusterDbRevisions request.}
}
\description{
Modifies the database revision of a cluster. The database revision is a unique revision of the database running in a cluster.
}
\examples{

}
