% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{modify_cluster_maintenance}
\alias{modify_cluster_maintenance}
\title{Modifies the maintenance settings of a cluster}
\usage{
modify_cluster_maintenance(ClusterIdentifier, DeferMaintenance = NULL,
  DeferMaintenanceIdentifier = NULL, DeferMaintenanceStartTime = NULL,
  DeferMaintenanceEndTime = NULL, DeferMaintenanceDuration = NULL)
}
\arguments{
\item{ClusterIdentifier}{[required] A unique identifier for the cluster.}

\item{DeferMaintenance}{A boolean indicating whether to enable the deferred maintenance window.}

\item{DeferMaintenanceIdentifier}{A unique identifier for the deferred maintenance window.}

\item{DeferMaintenanceStartTime}{A timestamp indicating the start time for the deferred maintenance window.}

\item{DeferMaintenanceEndTime}{A timestamp indicating end time for the deferred maintenance window. If you specify an end time, you can't specify a duration.}

\item{DeferMaintenanceDuration}{An integer indicating the duration of the maintenance window in days. If you specify a duration, you can't specify an end time. The duration must be 14 days or less.}
}
\description{
Modifies the maintenance settings of a cluster. For example, you can defer a maintenance window. You can also update or cancel a deferment.
}
\section{Accepted Parameters}{
\preformatted{modify_cluster_maintenance(
  ClusterIdentifier = "string",
  DeferMaintenance = TRUE|FALSE,
  DeferMaintenanceIdentifier = "string",
  DeferMaintenanceStartTime = as.POSIXct("2015-01-01"),
  DeferMaintenanceEndTime = as.POSIXct("2015-01-01"),
  DeferMaintenanceDuration = 123
)
}
}

