% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{modify_cluster_snapshot}
\alias{modify_cluster_snapshot}
\title{Modifies the settings for a snapshot}
\usage{
modify_cluster_snapshot(
  SnapshotIdentifier = "string",
  ManualSnapshotRetentionPeriod = 123,
  Force = TRUE|FALSE
)
}
\arguments{
\item{SnapshotIdentifier}{[required] The identifier of the snapshot whose setting you want to modify.}

\item{ManualSnapshotRetentionPeriod}{The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

If the manual snapshot falls outside of the new retention period, you can specify the force option to immediately delete the snapshot.

The value must be either -1 or an integer between 1 and 3,653.}

\item{Force}{A Boolean option to override an exception if the retention period has already passed.}
}
\description{
Modifies the settings for a snapshot.
}
\examples{

}
