% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{modify_cluster_snapshot_schedule}
\alias{modify_cluster_snapshot_schedule}
\title{Modifies a snapshot schedule for a cluster}
\usage{
modify_cluster_snapshot_schedule(
  ClusterIdentifier = "string",
  ScheduleIdentifier = "string",
  DisassociateSchedule = TRUE|FALSE
)
}
\arguments{
\item{ClusterIdentifier}{[required] A unique identifier for the cluster whose snapshot schedule you want to modify.}

\item{ScheduleIdentifier}{A unique alphanumeric identifier for the schedule that you want to associate with the cluster.}

\item{DisassociateSchedule}{A boolean to indicate whether to remove the assoiciation between the cluster and the schedule.}
}
\description{
Modifies a snapshot schedule for a cluster.
}
\examples{

}
