% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{modify_cluster_subnet_group}
\alias{modify_cluster_subnet_group}
\title{Modifies a cluster subnet group to include the specified list of VPC subnets}
\usage{
modify_cluster_subnet_group(ClusterSubnetGroupName, SubnetIds,
  Description = NULL)
}
\arguments{
\item{ClusterSubnetGroupName}{The name of the subnet group to be modified.}

\item{SubnetIds}{An array of VPC subnet IDs. A maximum of 20 subnets can be modified in a single request.}

\item{Description}{A text description of the subnet group to be modified.}
}
\description{
Modifies a cluster subnet group to include the specified list of VPC subnets. The operation replaces the existing list of subnets with the new list of subnets.
}
\examples{

}
