% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{modify_snapshot_copy_retention_period}
\alias{modify_snapshot_copy_retention_period}
\title{Modifies the number of days to retain automated snapshots in the destination region after they are copied from the source region}
\usage{
modify_snapshot_copy_retention_period(ClusterIdentifier, RetentionPeriod)
}
\arguments{
\item{ClusterIdentifier}{[required] The unique identifier of the cluster for which you want to change the retention period for automated snapshots that are copied to a destination region.

Constraints: Must be the valid name of an existing cluster that has cross-region snapshot copy enabled.}

\item{RetentionPeriod}{[required] The number of days to retain automated snapshots in the destination region after they are copied from the source region.

If you decrease the retention period for automated snapshots that are copied to a destination region, Amazon Redshift will delete any existing automated snapshots that were copied to the destination region and that fall outside of the new retention period.

Constraints: Must be at least 1 and no more than 35.}
}
\description{
Modifies the number of days to retain automated snapshots in the destination region after they are copied from the source region.
}
\section{Accepted Parameters}{
\preformatted{modify_snapshot_copy_retention_period(
  ClusterIdentifier = "string",
  RetentionPeriod = 123
)
}
}

