% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{modify_snapshot_schedule}
\alias{modify_snapshot_schedule}
\title{Modifies a snapshot schedule}
\usage{
modify_snapshot_schedule(ScheduleIdentifier, ScheduleDefinitions)
}
\arguments{
\item{ScheduleIdentifier}{[required] A unique alphanumeric identifier of the schedule to modify.}

\item{ScheduleDefinitions}{[required] An updated list of schedule definitions. A schedule definition is made up of schedule expressions, for example, "cron(30 12 *)" or "rate(12 hours)".}
}
\description{
Modifies a snapshot schedule. Any schedule associated with a cluster is modified asynchronously.
}
\section{Accepted Parameters}{
\preformatted{modify_snapshot_schedule(
  ScheduleIdentifier = "string",
  ScheduleDefinitions = list(
    "string"
  )
)
}
}

