% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_package.R
\docType{package}
\name{paws.redshift-package}
\alias{paws.redshift}
\alias{paws.redshift-package}
\title{paws.redshift: Amazon Redshift}
\description{
This is an interface reference for Amazon Redshift. It
contains documentation for one of the programming or command line
interfaces you can use to manage Amazon Redshift clusters. Note that
Amazon Redshift is asynchronous, which means that some interfaces may
require techniques, such as polling or asynchronous callback handlers,
to determine when a command has been applied. In this reference, the
parameter descriptions indicate whether a change is applied
immediately, on the next instance reboot, or during the next
maintenance window. For a summary of the Amazon Redshift cluster
management interfaces, go to [Using the Amazon Redshift Management
Interfaces](http://docs.aws.amazon.com/redshift/latest/mgmt/using-aws-sdk.html).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
