% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{purchase_reserved_node_offering}
\alias{purchase_reserved_node_offering}
\title{Allows you to purchase reserved nodes}
\usage{
purchase_reserved_node_offering(
  ReservedNodeOfferingId = "string",
  NodeCount = 123
)
}
\arguments{
\item{ReservedNodeOfferingId}{[required] The unique identifier of the reserved node offering you want to purchase.}

\item{NodeCount}{The number of reserved nodes that you want to purchase.

Default: \code{1}}
}
\description{
Allows you to purchase reserved nodes. Amazon Redshift offers a predefined set of reserved node offerings. You can purchase one or more of the offerings. You can call the DescribeReservedNodeOfferings API to obtain the available reserved node offerings. You can call this API by providing a specific reserved node offering and the number of nodes you want to reserve.
}
\details{
For more information about reserved node offerings, go to \href{http://docs.aws.amazon.com/redshift/latest/mgmt/purchase-reserved-node-instance.html}{Purchasing Reserved Nodes} in the \emph{Amazon Redshift Cluster Management Guide}.
}
\examples{

}
