% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{rotate_encryption_key}
\alias{rotate_encryption_key}
\title{Rotates the encryption keys for a cluster}
\usage{
rotate_encryption_key(
  ClusterIdentifier = "string"
)
}
\arguments{
\item{ClusterIdentifier}{[required] The unique identifier of the cluster that you want to rotate the encryption keys for.

Constraints: Must be the name of valid cluster that has encryption enabled.}
}
\description{
Rotates the encryption keys for a cluster.
}
\examples{

}
