% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{delete_snapshot_copy_grant}
\alias{delete_snapshot_copy_grant}
\title{Deletes the specified snapshot copy grant}
\usage{
delete_snapshot_copy_grant(SnapshotCopyGrantName)
}
\arguments{
\item{SnapshotCopyGrantName}{[required] The name of the snapshot copy grant to delete.}
}
\description{
Deletes the specified snapshot copy grant.
}
\section{Accepted Parameters}{
\preformatted{delete_snapshot_copy_grant(
  SnapshotCopyGrantName = "string"
)
}
}

