% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.redshift_operations.R
\name{modify_cluster_parameter_group}
\alias{modify_cluster_parameter_group}
\title{Modifies the parameters of a parameter group}
\usage{
modify_cluster_parameter_group(ParameterGroupName, Parameters)
}
\arguments{
\item{ParameterGroupName}{The name of the parameter group to be modified.}

\item{Parameters}{An array of parameters to be modified. A maximum of 20 parameters can be modified in a single request.

For each parameter to be modified, you must supply at least the parameter name and parameter value; other name-value pairs of the parameter are optional.

For the workload management (WLM) configuration, you must supply all the name-value pairs in the wlm\_json\_configuration parameter.}
}
\description{
Modifies the parameters of a parameter group.
}
\details{
For more information about parameters and parameter groups, go to \href{http://docs.aws.amazon.com/redshift/latest/mgmt/working-with-parameter-groups.html}{Amazon Redshift Parameter Groups} in the \emph{Amazon Redshift Cluster Management Guide}.
}
\examples{

}
